deployment/backend/requirements.txt
waitress
flask
numpy
pandas
tensorflow-cpu